Uplift TypeVisitableExt into rustc_type_ir
